PixelBuffer changes
PixelBuffer information was moved to it's separate substruct in CORE struct: now it contains the data about the current pixel buffer and it's size. What's more, you can pop and insert new pixelbuffers at runtime (it's not working, need to work this around). CMakeLists.txt in the /src/ folder was modified so it builds all the examples in this directory. I've added more exampels, and I'm thinking about making more and more in the future. Lastly, I've expanded the KeyCode list with new keys!